addressed most comments + directory
